requirements.txt
loguru==0.7.2
openai==0.27.8
python-dotenv==0.1.2
fastapi==0.110.0
uvicorn==0.24.0.post1
python-multipart==0.0.6
pre-commit==3.6.2
apscheduler==3.10.4